fix: move the global session listener to a place in widget tree where the localizations are available
